Clarify 10mm Pistol Manufacturer
A bit of background: The Colt 1911 .45 ACP Handgun was produced by multiple companies, including Ithaca.  That has been assigned to the standard Colt .45 here as well, under the in-universe company "Duketha" to match the Ithaca M37 Combat Shotgun taking the Duke Shotgun sprites.  However, for the 10mm "Delta Elite" variant, the in-universe equivalent of Colt itself, Volt, has been used and the model has been added to have it placed more as a variant than a completely new type of pistol.