Initial Commit to add actual Dogtag sprites for Summoning Talisman
Going along with the idea of anomalous/demonic/magical items, the Summoning Talisman always looked out of place being simply three helmets in the inventory.  Since it's supposed to look like the leftover gear of a fallen comrade, giving it a set of dogtags that themselves are the anomalous item in question makes more sense.