LANGUAGE Cleanup
Move Mk. ### Identifier in front of the Nicknames, as they're the pseudo-model of that weapon.
Volt lost the patent to the ZM7 Liberator Rifle to Heckler & Mok as per the HD Manual, so they've been rebranded.
The SMG was produced by H&M, which was missed originally.
The Hunter was produced by DiFranco, which was also missed.
The Slayer was produced by Cavalry Arms, which again was also missed.
The Assault Pack & Radtech Half-Pack have been renamed into the Lightweight Backpack and Rucksack respectively, as their sprites would indicate the Half-Pack is more flexible than the Assault Pack, even though they're both half the size of a standard Backpack.